backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2
pydantic>=2.6
python-dotenv>=1.0
uvicorn[standard]>=0.30
whitenoise>=6.7
djangorestframework>=3.15
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.4
pandas>=2.2
pyyaml>=6.0.2
openpyxl>=3.1
httpx>=0.27
